Complete Depth-Aware Resnet Backbone and first test of Open3D
Due to the changes in the Depth-Aware operation it is now possible
and feasible to clompletely run depth-aware operations in the resnet
backbone to include more depth information in all levels of the feature
maps. Also a promising looking package, Open3D was tested for point
cloud registration, which might make the use of the subprocesses
superfluous.