Changed Resnet-Backbone only use DA Ops for Conv Skip Connections
In the trails before, the Resnet backbone always used da convs for
evrey convolutional operation, which made it very slow and hindered
convergence. The backbone was now changed to only use the da
convolutions for skip connections that change the width and height
of the feature map. This should decrease runtime and allow
convergence