Finished Writing Unittests for the Pose-Estimation using Image Features
These tests include sanity checks for the shapes and comparative
calculations as well as visual comparisons. For now, it seems that the
pose estimation is without bugs. If the model does not converge now,
there might be an error in the calculation of the gradient of the
next nearest neighbor tf op, which seems a little unaccurate. An
improvement could be something like
 loss = min(abs(gt_pose-pred_pose), chamfer_loss)