Fixed 2 Bugs in the pose estimation
The matrices returned by the pose estimation branch are not orthogonal,
therefore an svd is now used to make them orthogonal for the loss
calculation.

It is also not possible to simply transform a padding into camera coords
and then add this as an offset to the objects poses. Now, the
translational part of the pose is transformed to pixels, then the
offset is added and everything is back transformed.