Added Translational Loss Term for the Pose Estimation
Apparently, using only the chamfer distance is not enough for the
network to learn the correct pose. Therefore, the MAE for the
rotational matrices and the translation vectors is computed seperately
and added to the chamfer_loss. This should help the network learn
independently of the chamfer losses implementation (which might be
buggy)