Changes to the PointNet T-Graph
Before there were different weights per class for the convolutional
layers in the T-Graph and different weights per class and roi for
the fully connected layers. This might be a bit overkill, and maybe
even wrong as the weights should probaly not differ for the rois.

Now the weights are the same for all rois and classes for the
convolutional layers and differ only for the classes in the fc
layers.